+++ pinot-common/src/test/java/org/apache/pinot/util/TestUtils.java
@@ -26,10 +26,12 @@ import java.time.Duration;
 import java.time.Instant;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import net.openhft.chronicle.core.Jvm;
 import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testng.Assert;
+import org.testng.SkipException;
 
 
 /**
@@ -198,6 +200,12 @@ public class TestUtils {
     }
 
     throw new IllegalStateException("Failed to return result in " + timeoutMs + "ms");
+  }
+
+  public static void ensureArchitectureIsNotARM() {
+    if (Jvm.isArm()) {
+      throw new SkipException("Skipping test as the underlying architecture is ARM");
+    }
   }
 
   @FunctionalInterface
