+++ pinot-segment-local/src/test/java/org/apache/pinot/segment/local/segment/index/creator/CLPForwardIndexCreatorTest.java
@@ -51,6 +51,9 @@ public class CLPForwardIndexCreatorTest {
   @BeforeClass
   public void setUp()
       throws Exception {
+    //skip this test if the underlying arch is aarch64 as CLP isn't supported on ARM yet:
+    // https://github.com/y-scope/clp-ffi-java/issues/46
+    TestUtils.ensureArchitectureIsNotARM();
     TestUtils.ensureDirectoriesExistAndEmpty(TEMP_DIR);
   }
